All resource tests are in one file named ResourceTest.php
